requirements.txt
Python 3.8+
OpenCV
PyTorch
DeepFace
Ultralytics #YOLO
NumPy
CUDA-capable # GPU (recommended for faster processing)